Add RequestHandler for Urban Dictionary requests.
Just made a basic Urban Dictionary request handler which takes in Python's requests object as a data member
and proceeds to make the requests and extract the desired information from it. Ultimately, it makes a get
request on a given term, abstracts the desired information, and proceeds to validate it.