refactor to naming conventions for the request handler
Changed the name of main.py -> client.py and UDRequestHandler.py -> UDRequest since
the object is not actually recieving a request itself. Also fixed the __call__ dunder
method. It was not returning the value, which broke the client.py function when using
it for the refactor. Also the UDRequest object now takes in no arguments into the
constructor and uses the requests get metod itself.