Set up simple server
Added support for tornado, along with a simple UDRequestHandler which just sends a UDRequest
to the Urban Dictionary API, processes the result, and writes it as a http response.
Added Docker support for a live server that current runs on port 5127 with no config file.
Removed the live shell, as it's not currently working/relevant. May need to readd this
in the future while running the python server in the background.